+++ src/unsigned_int.rs
@@ -3,7 +3,7 @@ use crate::{False, Integer, IsNonZero, IsSigned, NonZero, SignedInt, Splat};
 /// Unsigned integer common operations.
 #[allow(clippy::len_without_is_empty)]
 pub trait UnsignedInt:
-    IsSigned<Signed = False> + IsNonZero<NonZero = False> + Integer + Splat<u8>
+    IsSigned<Signed = False> + IsNonZero<NonZero = False> + Integer<Unsigned = Self> + Splat<u8>
 {
     /// The signed variant of [`UnsignedInt`].
     type SignedInt: SignedInt<UnsignedInt = Self>;